Make Rake platform configurable
Since commit 30dce5b, `rake-build` errors out when running it under
Windows with Ruby installed in a Unix-style subsystem like Babun or
Cygwin. The error happens because the plugin currently expects the
Rake executable to be named `rake.bat`, while it’s actually `rake`
even though it’s on Windows.

This PR adds a configuration setting called `rakePlatform`, which
defaults to the same auto-detection that was already implemented. It
can be changed either via `config.cson` or the Settings pane.